update: commandHandler was refactorized to be clean and increase its maintenability
